+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.x] - 2025-06-16
 
-## [0.1.0] - 2024-12-19
+### Added
+- Add `--registry-url` flag to specify the Schema Registry instance URL
+- Add `--user` flag to specify the username for authentication
+- Add `--pass` flag to specify the password for authentication
+- Add `--api-key` flag to specify the API key for authentication
+
+## [0.1.x] - 2025-06-15
 
 ### Added
 - Initial release of ksr-cli